# backend/docs/swagger.yaml
basePath: /
definitions:
...
  web-hosting_internal_modules_kelas_dto.JurusanResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  web-hosting_internal_modules_kelas_dto.KelasCreateRequest:
    properties:
      kurikulum_kode:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        type: string
      prodi_name:
        maxLength: 255
        type: string
      semester:
        minimum: 0
        type: integer
      tahun_akademik_id:
        type: integer
    required:
    - kurikulum_kode
    - name
    - prodi_name
    - tahun_akademik_id
    type: object
...
  web-hosting_internal_modules_kelas_dto.KelasResponse:
    properties:
      id:
        type: string
      kurikulum:
        $ref: '#/definitions/web-hosting_internal_modules_kelas_dto.KurikulumResponse'
      mahasiswa:
        items:
          $ref: '#/definitions/web-hosting_internal_modules_kelas_dto.MahasiswaResponse'
        type: array
      name:
        type: string
      prodi:
        $ref: '#/definitions/web-hosting_internal_modules_kelas_dto.ProdiResponse'
      semester:
        type: integer
      tahun_akademik:
        $ref: '#/definitions/web-hosting_internal_modules_kelas_dto.TahunAkademikResponse'
    type: object
...
  web-hosting_internal_modules_kelas_dto.KelasUpdateRequest:
    properties:
      kurikulum_kode:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        type: string
      prodi_name:
        maxLength: 255
        type: string
      semester:
        minimum: 0
        type: integer
      tahun_akademik_id:
        type: integer
    type: object
  web-hosting_internal_modules_kelas_dto.KurikulumMKResponse:
    properties:
      mata_kuliah:
        $ref: '#/definitions/web-hosting_internal_modules_kelas_dto.MataKuliahResponse'
      semester:
        type: integer
      wajib:
        type: boolean
    type: object
  web-hosting_internal_modules_kelas_dto.KurikulumResponse:
    properties:
      kode:
        type: string
      kurikulum_mk:
        items:
          $ref: '#/definitions/web-hosting_internal_modules_kelas_dto.KurikulumMKResponse'
        type: array
      name:
        type: string
    type: object
  web-hosting_internal_modules_kelas_dto.MahasiswaResponse:
    properties:
      email:
        type: string
      mahasiswa_id:
        type: string
      name:
        type: string
    type: object
  web-hosting_internal_modules_kelas_dto.MataKuliahResponse:
    properties:
      kode:
        type: string
      name:
        type: string
      sks:
        type: integer
    type: object
  web-hosting_internal_modules_kelas_dto.ProdiResponse:
    properties:
      id:
        type: integer
      jenjang:
        type: string
      jurusan:
        $ref: '#/definitions/web-hosting_internal_modules_kelas_dto.JurusanResponse'
      name:
        type: string
    type: object
  web-hosting_internal_modules_kelas_dto.TahunAkademikResponse:
    properties:
      id:
        type: integer
      status:
        type: string
      tahun_akhir:
        type: string
      tahun_awal:
        type: string
      tipe_semester:
        type: string
    type: object
...
  web-hosting_internal_package_swagger.ErrCreateKelasFailed:
    properties:
      error:
        example: kelas already exists
        type: string
      message:
        example: failed to create kelas
        type: string
      path:
        example: /api/kelas
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrCreateKelasInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to create kelas
        type: string
      path:
        example: /api/kelas
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrDeleteKelasFailed:
    properties:
      error:
        example: kelas already exists
        type: string
      message:
        example: failed to delete kelas
        type: string
      path:
        example: /api/kelas/{kelas_id}
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrDeleteKelasInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to delete kelas
        type: string
      path:
        example: /api/kelas/{kelas_id}
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrForbiddenAccess:
    properties:
      error:
        example: Forbidden
        type: string
      message:
        example: Role anda tidak diizinkan
        type: string
      path:
        example: /api/super/role
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrGetKelasFailed:
    properties:
      error:
        example: kelas not found
        type: string
      message:
        example: failed to get kelas
        type: string
      path:
        example: /api/kelas/{kelas_id}
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrGetKelasInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to get kelas
        type: string
      path:
        example: /api/kelas/{kelas_id}
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken:
    properties:
      error:
        example: invalid token
        type: string
      message:
        example: failed_auth
        type: string
      path:
        example: /api/auth/logout
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrUpdateKelasFailed:
    properties:
      error:
        example: kelas already exists
        type: string
      message:
        example: failed to update kelas
        type: string
      path:
        example: /api/kelas/{kelas_id}
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrUpdateKelasInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to update kelas
        type: string
      path:
        example: /api/kelas/{kelas_id}
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_utils.PaginatedData-array_web-hosting_internal_modules_kelas_dto_KelasResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/web-hosting_internal_modules_kelas_dto.KelasResponse'
        type: array
      pagination:
        $ref: '#/definitions/web-hosting_internal_package_utils.PaginationMeta'
    type: object
...
  web-hosting_internal_package_utils.PaginationMeta:
    properties:
      page:
        example: 1
        type: integer
      per_page:
        example: 10
        type: integer
      total_items:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  web-hosting_internal_package_utils.Response-any-any:
    properties:
      data: {}
      error: {}
      message:
        example: Operation successful
        type: string
      path:
        example: /api/resource
        type: string
      success:
        type: boolean
    type: object
...
  web-hosting_internal_package_utils.Response-web-hosting_internal_modules_kelas_dto_KelasResponse-any:
    properties:
      data:
        $ref: '#/definitions/web-hosting_internal_modules_kelas_dto.KelasResponse'
      error: {}
      message:
        example: Operation successful
        type: string
      path:
        example: /api/resource
        type: string
      success:
        type: boolean
    type: object
...
  ? web-hosting_internal_package_utils.Response-web-hosting_internal_package_utils_PaginatedData-array_web-hosting_internal_modules_kelas_dto_KelasResponse-any
  : properties:
      data:
        $ref: '#/definitions/web-hosting_internal_package_utils.PaginatedData-array_web-hosting_internal_modules_kelas_dto_KelasResponse'
      error: {}
      message:
        example: Operation successful
        type: string
      path:
        example: /api/resource
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    **BARANG SIAPA YANG UPDATE/HAPUS DATA SEMBARANGAN DAPAT KECUPAN DARI HAFIZ**.
    **BUAT DATA BARU JIKA INGIN HAPUS/UPDATE**.

    **DUMMY DATA LINK:**
    **https://github.com/Kar-Su/pbl-project-simpadu/tree/main/backend/internal/database/seeders/json**

    **TUTORIAL LINK:***
    **https://github.com/Kar-Su/pbl-project-simpadu**
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TIM 1 API
  version: "1.0"
paths:
...
  /api/kelas:
    post:
      consumes:
      - application/json
      description: |-
        Menambahkan kelas baru ke sistem

        **Akses:** Admin Akademik

        **Error yang mungkin terjadi:**
        - `400` Body tidak valid / field wajib kosong -> `message: "failed to get request", error: "Key: 'KelasName' Error:..."`
        - `400` kelas dengan id tersebut sudah ada -> `message: "failed to create kelas", error: "kelas already exists"`
        - `401` Authorization header tidak ada -> `message: "failed_auth", error: "Authorization header missing"`
        - `401` Format header salah (bukan "Bearer ...") -> `message: "failed_auth", error: "invalid authentication header"`
        - `401` Token JWT tidak valid atau kedaluwarsa -> `message: "failed_auth", error: "invalid token"`
        - `403` user tidak memiliki akses -> `message: "kelas anda tidak diizinkan", error: "Forbidden"`
        - `500` Kesalahan internal server -> `message: "failed to create kelas", error: "Internal Error"`
      parameters:
      - description: Kelas Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web-hosting_internal_modules_kelas_dto.KelasCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-any-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrCreateKelasFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrCreateKelasInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Create Kelas Baru
      tags:
      - kelas
  /api/kelas/{kelas_id}:
    delete:
      consumes:
      - application/json
      description: |-
        delete kelas yang sudah ada

        **Akses:** Admin Akademik

        **Error yang mungkin terjadi:**
        - `400` Parameter tidak valid -> `message: "failed to validate parameter", error: "Key: 'param' Error:..."`
        - `400` kelas dengan id tersebut tidak ditemukan -> `message: "failed to Delete kelas", error: "kelas not found"`
        - `401` Authorization header tidak ada -> `message: "failed_auth", error: "Authorization header missing"`
        - `401` Format header salah (bukan "Bearer ...") -> `message: "failed_auth", error: "invalid authentication header"`
        - `401` Token JWT tidak valid atau kedaluwarsa -> `message: "failed_auth", error: "invalid token"`
        - `403` kelas user tidak memiliki akses -> `message: "kelas anda tidak diizinkan", error: "Forbidden"`
        - `500` Kesalahan internal server -> `message: "failed to Delete kelas", error: "Internal Error"`
      parameters:
      - description: kelas ID
        in: path
        name: kelas_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-any-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrDeleteKelasFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrDeleteKelasInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Delete kelas
      tags:
      - kelas
    get:
      consumes:
      - application/json
      description: |-
        melihat kelas yang sudah ada

        **Akses:** Logged User

        **Error yang mungkin terjadi:**
        - `400` Parameter tidak valid -> `message: "failed to validate parameter", error: "Key: 'param' Error:..."`
        - `400` kelas dengan id tersebut tidak ditemukan -> `message: "failed to Get kelas", error: "kelas not found"`
        - `401` Authorization header tidak ada -> `message: "failed_auth", error: "Authorization header missing"`
        - `401` Format header salah (bukan "Bearer ...") -> `message: "failed_auth", error: "invalid authentication header"`
        - `401` Token JWT tidak valid atau kedaluwarsa -> `message: "failed_auth", error: "invalid token"`
        - `403` kelas user tidak memiliki akses -> `message: "kelas anda tidak diizinkan", error: "Forbidden"`
        - `500` Kesalahan internal server -> `message: "failed to Get kelas", error: "Internal Error"`
      parameters:
      - description: kelas ID
        in: path
        name: kelas_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-web-hosting_internal_modules_kelas_dto_KelasResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrGetKelasFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrGetKelasInternalServer'
      security:
      - ApiKeyAuth: []
      summary: get Kelas
      tags:
      - kelas
    put:
      consumes:
      - application/json
      description: |-
        Mengupdate kelas yang sudah ada

        **Akses:** Admin Akademik

        **Error yang mungkin terjadi:**
        - `400` Parameter tidak valid -> `message: "failed to validate parameter", error: "Key: 'param' Error:..."`
        - `400` Body tidak valid / field wajib kosong -> `message: "failed to get request", error: "Key: 'Kelas' Error:..."`
        - `400` kelas dengan id tersebut tidak ditemukan -> `message: "failed to update kelas", error: "kelas not found"`
        - `401` Authorization header tidak ada -> `message: "failed_auth", error: "Authorization header missing"`
        - `401` Format header salah (bukan "Bearer ...") -> `message: "failed_auth", error: "invalid authentication header"`
        - `401` Token JWT tidak valid atau kedaluwarsa -> `message: "failed_auth", error: "invalid token"`
        - `403` role user tidak memiliki akses -> `message: "role anda tidak diizinkan", error: "Forbidden"`
        - `500` Kesalahan internal server -> `message: "failed to update role", error: "Internal Error"`
      parameters:
      - description: kelas ID
        in: path
        name: kelas_id
        required: true
        type: string
      - description: kelas Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web-hosting_internal_modules_kelas_dto.KelasUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-web-hosting_internal_modules_kelas_dto_KelasResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUpdateKelasFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUpdateKelasInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Update kelas
      tags:
      - kelas
...
  /api/kelas/prodi/{prodi_name}:
    get:
      consumes:
      - application/json
      description: |-
        Mendapatkan daftar kelas berdasarkan nama prodi, dengan pagination (10 per halaman).
        Setiap kelas sudah termasuk deep join: kurikulum → kurikulum_mk (difilter by semester kelas) → mata_kuliah, prodi → jurusan, tahun_akademik, dan daftar mahasiswa.

        **Akses:** Logged User

        **Error yang mungkin terjadi:**
        - `400` Parameter URI tidak valid -> `message: "failed to Get kelas", error: "Key: 'param' Error:..."`
        - `400` Prodi tidak ditemukan -> `message: "failed to Get kelas", error: "prodi not found"`
        - `401` Authorization header tidak ada -> `message: "failed_auth", error: "Authorization header missing"`
        - `401` Format header salah (bukan "Bearer ...") -> `message: "failed_auth", error: "invalid authentication header"`
        - `401` Token JWT tidak valid atau kedaluwarsa -> `message: "failed_auth", error: "invalid token"`
        - `500` Kesalahan internal server -> `message: "failed to Get kelas", error: "Internal Error"`
      parameters:
      - description: Nama Prodi
        example: teknik-listrik
        in: path
        name: prodi_name
        required: true
        type: string
      - description: Halaman (default 1, 10 per halaman)
        example: 1
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-web-hosting_internal_package_utils_PaginatedData-array_web-hosting_internal_modules_kelas_dto_KelasResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrGetKelasFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrGetKelasInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Get Kelas by Prodi (Paginated)
      tags:
      - kelas
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Format: Bearer <access token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"